Update socket flow and Refactor
1.
- In back-end index socket, add 2 emitters: updatedAllUsersFlag, updatedRankingListFlag
- Add listeners on front-end Layout

2. 
- Refactor into a folder redis-utils: RedisUtil, TransactionsHistoryMost5RecentlyUsed, SharesInfoBank
