+++ front-end/src/App.js
@@ -38,11 +38,6 @@ var socket;
 const BACKEND_HOST_FOR_SOCKET = getBackendHostForSocket();
 
 class App extends React.Component {
-  constructor(props) {
-    super(props);
-    socket = socketIOClient(BACKEND_HOST_FOR_SOCKET);
-  }
-
   // variables
   specialLinks = [
     "/login",
@@ -99,6 +94,7 @@ class App extends React.Component {
   };
 
   componentDidMount() {
+    socket = socketIOClient(BACKEND_HOST_FOR_SOCKET);
     this.changePath();
     this.setupReduxStoreState();
   }
